allow deletion of reviews

+++ backend/routes/reviewRatings.js
@@ -30,9 +30,17 @@ router.get('/:mediaType/:mediaId', (req, res) => {
 });
 
 router.put('/review', (req, res) => {
-    const { text, mediaId, mediaType, datePosted } = req.body;
-    ReviewRating.findOneAndUpdate({ mediaType: mediaType, id: mediaId },
-        { mediaType: mediaType, id: mediaId, text: text, datePosted: datePosted },
+    const { mediaId, mediaType, datePosted, text, userName, userId } = req.body;
+    ReviewRating.findOneAndUpdate(
+        { mediaType: mediaType, id: mediaId, userId: userId },
+        {
+                mediaType: mediaType,
+                id: mediaId,
+                text: text,
+                datePosted: datePosted,
+                userName: userName,
+                userId: userId
+            },
         { new: true, upsert: true })
         .then(reviewRating => {
             console.log('Success. Updated or posted review: ', reviewRating);
@@ -45,9 +53,9 @@ router.put('/review', (req, res) => {
 });
 
 router.put('/rating', (req, res) => {
-    const { rating, mediaId, mediaType } = req.body;
-    ReviewRating.findOneAndUpdate({ mediaType: mediaType, id: mediaId },
-        { mediaType: mediaType, id: mediaId, rating: rating },
+    const { rating, userId, mediaId, mediaType } = req.body;
+    ReviewRating.findOneAndUpdate({ mediaType: mediaType, id: mediaId, userId: userId },
+        { mediaType: mediaType, userId: userId, id: mediaId, rating: rating },
         { new: true, upsert: true })
         .then(reviewRating => {
             console.log('Success. Updated or posted rating: ', reviewRating);
@@ -59,9 +67,9 @@ router.put('/rating', (req, res) => {
         });
 });
 
-router.delete('/:mediaType/:mediaId', (req, res) => {
-    const { mediaId, mediaType } = req.params;
-    ReviewRating.deleteOne({ mediaType: mediaType, id: mediaId })
+router.delete('/:mediaType/:mediaId/:userId', (req, res) => {
+    const { mediaId, mediaType, userId } = req.params;
+    ReviewRating.deleteOne({ mediaType: mediaType, id: mediaId, userId: userId })
         .then(result => {
             console.log('Success. Deleted reviewRating document, count:', result.deletedCount);
             res.status(200).json(result);
